// web-portal/back-center-vue/src/common/utils/permission/permission-base-enum.ts
export enum PermissionBaseEnum {
  ADD = 'ADD',
  MODIFY = 'MODIFY',
  DELETE = 'DELETE',
  QUERY_PAGED = 'QUERY::PAGED',
  QUERY_LIST = 'QUERY::LIST',
  QUERY_ID = 'QUERY::ID',
  UPDATE_ENABLED = 'UPDATE::ENABLED',
  USER_RESET_PASSWORD = 'USER::RESET::PASSWORD',
  UPLOAD = 'UPLOAD',
  DOWNLOAD = 'DOWNLOAD',
  IMPORT = 'IMPORT',
  EXPORT = 'EXPORT'
}